match the real 4.4BSD include file for recv, recvfrom, recvmsg, send, sendto, sedmsg
SCCS-vsn: contrib/libg++-2.3/libg++/g++-include/sys/socket.h 1.3

+++ usr/src/contrib/libg++-2.3/libg++/g++-include/sys/socket.h
@@ -31,14 +31,14 @@ int       listen(int, int);
 #ifndef hpux
 int       rcmd _G_ARGS((char**, int, const char*, const char*, const char*, int*));
 #endif
-int       recv(int, void*, int, int);
-int       recvfrom _G_ARGS((int, void*, int, int, struct sockaddr *, int *));
-int       recvmsg(int, struct msghdr*, int);
+ssize_t	  recv(int, void*, int, int);
+ssize_t	  recvfrom _G_ARGS((int, void*, int, int, struct sockaddr *, int *));
+ssize_t	  recvmsg(int, struct msghdr*, int);
 int       rexec(char**, int, const char*, const char*, const char*, int*);
 int       rresvport(int*);
-int       send _G_ARGS((int, const void*, int, int));
-int       sendto _G_ARGS((int, const void*, int, int, const struct sockaddr*, int));
-int       sendmsg _G_ARGS((int, const struct msghdr*, int));
+ssize_t	  send _G_ARGS((int, const void*, int, int));
+ssize_t	  sendto _G_ARGS((int, const void*, int, int, const struct sockaddr*, int));
+ssize_t	  sendmsg _G_ARGS((int, const struct msghdr*, int));
 int       setsockopt _G_ARGS((int, int, int, const void *, int));
 int       shutdown(int, int);
 int       socket(int, int, int);
